+++ libs/ngx-editorjs/src/index.ts
@@ -6,9 +6,16 @@ export { Block, HeaderBlock, ImageBlock, ListBlock, ListStyle, ParagraphBlock } 
 export { EditorJSConfig, NgxEditorJSConfig, NGX_EDITORJS_CONFIG, EDITIOR_JS_INSTANCE } from './lib/types/config';
 export { BasePlugin, UserPlugins, PluginConfig } from './lib/types/plugins';
 export { createEditorJSConfig } from './lib/config/editor-config';
-export { EditorJSContainerComponent } from './lib/containers/base/container.class';
-export { EditorJSFormField, EDITORJS_FORM_VALUE_ACCESSOR } from './lib/containers/base/form-field.class';
+export { EditorJSBaseComponent, EDITORJS_FORM_VALUE_ACCESSOR } from './lib/containers/base/container.class';
 export {
   NgxEditorJSMatFieldComponent,
   EDITORJS_MATERIAL_FIELD_CONTROL
 } from './lib/containers/editorjs-mat-field/editorjs-mat-field.component';
+
+export {
+  NgxEditorJSInstanceService as ɵNgxEditorJSInstanceService,
+  EDITORJS_MODULE_IMPORT as ɵEDITORJS_MODULE_IMPORT,
+  EditorJSInstanceConfig as ɵEditorJSInstanceConfig
+} from './lib/services/editorjs-injector';
+
+export { NgxEditorJSPluginService as ɵNgxEditorJSPluginService } from './lib/services/plugins.service';